Custom functions don't have to specify fit.
Some custom functions do not need to be fit before they produce data. Added some
logic to enable not having to specify a dummy fit function. An empty string
after 'fit' in JSON indicates that no fitting is to be done for a step.

TESTED=Applied this to traditional_image and the test passes.